SVG: apply output size string to the actual page width and height
When the client supplied a geometry string as in 'gm convert
-size "50x50%" in.svg out.png', the output image columns and rows
would be incorrectly set from the size string to (in this example)
50.  Later the 50% scale factor would be applied, resulting in a
25 x 25 output image.  The initial setting of the output image
columns and rows is now zeroed out so that they will be correctly
set from either the svg "canvas" width/height, or from the svg
viewbox.  The geometry string is then later correctly applied to
the actual output image dimensions.