SVG: Restore previous behavior for non-SVG clippping paths (+ other).
Change set 9aaeeca0224c modified the drawing of clipping
paths to conform to the SVG spec.  This change set restores
the previous behavior for non-SVG clients of render.c, while
still satisfying the SVG spec for SVG clients.

For SVG files the initial value of "fill-rule" is now set to
"nonzero" (the SVG default) instead of the internally initialized
value of "evenodd".

C99 "//" comments were fixed in coders/wpg.c