SVG: Fix write beyond array bounds in TraceStrokePolygon.
Excessively large values of stroke-width were cascading
through other computations, causing TraceStrokePolygon()
in render.c to write beyond the end of it's array of points
when the stroke-linejoin attribute value was "round". Code
was added to reallocate the array of points as needed, and
to limit the size of stroke-width (for computational
purposes) to no more than approximately twice the diagonal
size of the output image.  Fixes ticket #515.

The same limit on stroke-width was applied to all other
instances of the same computation in render.c.