SVG: Fix ticket #516 TraceEllipse Heap Out of Bounds write.
The PrimitiveInfo array used to store points generated by
TraceEllipse(), the other TraceXXX() functions, and
DrawImage() was not always being expanded when needed,
resulting in writes beyond the end of the currently
allocated storage. To fix this problem, a new data structure
PrimitiveInfoMgr, and an associated function,
PrimtiveInfoRealloc(), were written to handle expanding the
PrimitiveInfo array as needed. DrawImage() and the TraceXXX()
functions were modified to use PrimitiveInfoRealloc() to
prevent the out of bounds writes to memory. This fixes ticket
#516.